requirements.txt
folium
pathlib
os
random
shutil
pillow
math
exifread
datetime
flask 
functools
pyfladesk
sys
pickle